report-worker/requirements.txt
mysql-connector-python
python-docx
requests
